requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.7.1
dill==0.3.7
Django==4.2.6
django-allauth==0.57.0
django-apscheduler==0.6.2
django-filter==23.3
django-modeltranslation==0.18.11
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
idna==3.4
isort==5.13.2
kombu==5.3.3
mccabe==0.7.0
oauthlib==3.2.2
pip==24.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.1.0
PyJWT==2.8.0
pylint==3.0.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.10